Push the ev dependency from fmidi-player to fmidi-play
The fmidi player library (in `fmidi_player.cc`) actually does not depend on libev. This patch adds an option to eventually disable the player library, and moves the dependency checker for libev to the `fmidi-play` program. Given that the header is the same, maybe the fmidi-player library should just be merged with the fmidi library?